requirements.txt
PyYAML==6.0.2
streamlit==1.38.0
streamlit_authenticator
loguru==0.7.3
